requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.1.2
black==24.8.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.10.1
cryptography==45.0.6
distro==1.9.0
flake8==7.1.1
Flask==3.0.0
Flask-Cors==5.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.2
openai==1.99.9
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.2.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.3
pytest==8.3.2
pytest-cov==6.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
rapidfuzz==3.6.1
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
Werkzeug==3.0.3
